Mac: Remove failing explicit conversion to nint in HatchedBar (and RunningDots), that caused only partial animation of the bar.

+++ FormsTest/Controls/RunningDots.cs
@@ -43,7 +43,7 @@ namespace FormsTest
 			replicatorLayer = new CAReplicatorLayer();
 			replicatorLayer.Frame = new CGRect(0, 0, view.Bounds.Width*2, view.Bounds.Height);
 			replicatorLayer.PreservesDepth = false;
-			replicatorLayer.InstanceCount = (nint)shapeCount;
+			replicatorLayer.InstanceCount = (int)shapeCount;
 			replicatorLayer.InstanceTransform = CATransform3D.MakeTranslation(distance, 0, 0);
 			replicatorLayer.InstanceDelay = -0.06f;
 			if (view.Layer == null)
